Revert "feat: add api route"
